implementando orm prisma con RDS

+++ src/controllers/ProfesorController.ts
@@ -10,61 +10,87 @@ export class ProfesorController {
     ){
   }
   
-  public getAllProfesores(req: Request, res: Response){
-    const profesores = this.profesorRepository.getAllProfesores();
-    res.status(200).send(profesores); 
-  }
-
-  public getProfesorById(id: number, req: Request, res: Response){
-    const profesor = this.profesorRepository.getProfesorById(id);
-  
-    if (profesor) {
-      res.status(200).send(profesor); 
-    } else {
-      res.status(404).send({ error: 'Profesor no encontrado'});
+  public async getAllProfesores(req: Request, res: Response){
+    try {
+      const profesores = await this.profesorRepository.getAllProfesores();
+      res.status(200).json(profesores);
+    } catch (error) {
+      res.status(500).json({ error: 'Unable to fetch items' });
     }
   }
 
-  public createProfesor(req: Request, res: Response) {
-    const { id, numeroEmpleado, nombres, apellidos, horasClase } = req.body;
+  public async getProfesorById(id: number, req: Request, res: Response){
+  try {
+    const profesor = await this.profesorRepository.getProfesorById(id);
+    if(profesor) {
+      res.status(200).send(profesor); 
+    }else{
+      res.status(404).send({ error: 'Profesor no encontrado'});
+    }
+  } catch (error) {
+    res.status(500).json({ error: 'Unable to fetch item' });
+  }
+
+  }
+
+  public async createProfesor(req: Request, res: Response) {
+    const { numeroEmpleado, nombres, apellidos, horasClase } = req.body;
     
-    const nuevoProfesor = new Profesor(id, numeroEmpleado, nombres, apellidos, horasClase);
+    const nuevoProfesor = new Profesor( numeroEmpleado, nombres, apellidos, horasClase);
   
     if (!this.validateProfesor.isValid(nuevoProfesor)) {
       res.status(400).send({ error: 'Datos de profesor no válidos'});
       return;
     }
+    
+    try {
+      const profesor = await this.profesorRepository.addProfesor(nuevoProfesor);
+      const id = profesor.id;
+      res.status(201).send({ message: 'Profesor creado exitosamente',id});
+    } catch (error) {
+      res.status(500).send({ error: 'Error al crear el alumno' });
+    }
   
-    this.profesorRepository.addProfesor(nuevoProfesor);
-    res.status(201).send({ message: 'Profesor creado exitosamente'});
   }
 
-  public updateProfesor(id: number, req: Request, res: Response){
+  public async updateProfesor(id: number, req: Request, res: Response){
     const { numeroEmpleado, nombres, apellidos, horasClase } = req.body;
   
-    const updatedProfesor = new Profesor(id, numeroEmpleado, nombres, apellidos, horasClase);
+    const updatedProfesor = new Profesor( numeroEmpleado, nombres, apellidos, horasClase);
   
     if (!this.validateProfesor.isValid(updatedProfesor)) {
       res.status(400).send({ error: 'Datos de profesor no válidos'}); 
       return;
     }
   
-    const updated = this.profesorRepository.updateProfesor(updatedProfesor);
-  
-    if (updated) {
-      res.status(200).send({ message: 'Profesor actualizado exitosamente'}); 
-    } else {
-      res.status(404).send({ error: 'Profesor no encontrado para actualizar'});
+    try {
+      const profesorExistente = await this.profesorRepository.getProfesorById(id)
+      if(!profesorExistente){
+        res.status(404).send({ error: 'Profesor no encontrado'})
+        return;
+      }
+
+      const updated = await this.profesorRepository.updateProfesor(id, updatedProfesor);
+      res.status(200).send({ message: 'Profesor actualizado exitosamente'});
+
+    } catch (error) {
+      res.status(500).send({ error: 'Error servidor' });
     }
+  
   }
 
-  public deleteProfesor(id: number,req: Request, res: Response){
-    const deleted = this.profesorRepository.deleteProfesor(id);
-  
-    if (deleted) {
+  public async deleteProfesor(id: number,req: Request, res: Response){
+try {
+  const profesorExistente = await this.profesorRepository.getProfesorById(id)
+      if(!profesorExistente){
+        res.status(404).send({ error: 'Profesor no encontrado'})
+        return;
+      }
+      const deleted = this.profesorRepository.deleteProfesor(id);
       res.status(200).send({ message: 'Profesor eliminado exitosamente'});
-    } else {
-      res.status(404).send({ error: 'Profesor no encontrado para eliminar'});
-    }
+} catch (error) {
+  res.status(500).send({error: 'Error servidor'});
+}
+
   }
 }
